git-svn-id: svn://123.57.173.36/meijia-parent@521 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/user/UserFriends.java
@@ -7,7 +7,7 @@ public class UserFriends {
 
     private Long friendId;
 
-    private Integer addTime;
+    private Long addTime;
 
     private Long updateTime;
 
@@ -35,11 +35,11 @@ public class UserFriends {
         this.friendId = friendId;
     }
 
-    public Integer getAddTime() {
+    public Long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(Integer addTime) {
+    public void setAddTime(Long addTime) {
         this.addTime = addTime;
     }
